+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Ability for Skills to write to Spaces (using External Nodes).
 - Add API support for creating and managing PaperQA collections.
 
+### Fixed
+
+- Improve admin performance by not loading all objects for FK dropdowns.
+
 ## [25.5.2] - 2025-05-19
 
 ### Fixed